:bug: Bugfix for string annotations
